Extract PyBind11 feature implementations
Differential Revision: D30678404

fbshipit-source-id: 2922aedfa1242823f8c39813245585d55be97b39
