+++ torchaudio/csrc/sox/io.h
@@ -1,15 +1,15 @@
 #ifndef TORCHAUDIO_SOX_IO_H
 #define TORCHAUDIO_SOX_IO_H
 
-#ifdef TORCH_API_INCLUDE_EXTENSION_H
-#include <torch/extension.h>
-#endif // TORCH_API_INCLUDE_EXTENSION_H
-
 #include <torch/script.h>
 #include <torchaudio/csrc/sox/utils.h>
 
 namespace torchaudio {
 namespace sox_io {
+
+std::vector<std::vector<std::string>> get_effects(
+    const c10::optional<int64_t>& frame_offset,
+    const c10::optional<int64_t>& num_frames);
 
 std::tuple<int64_t, int64_t, int64_t, int64_t, std::string> get_info_file(
     const std::string& path,
@@ -32,32 +32,6 @@ void save_audio_file(
     c10::optional<std::string> format,
     c10::optional<std::string> encoding,
     c10::optional<int64_t> bits_per_sample);
-
-#ifdef TORCH_API_INCLUDE_EXTENSION_H
-
-std::tuple<int64_t, int64_t, int64_t, int64_t, std::string> get_info_fileobj(
-    py::object fileobj,
-    c10::optional<std::string> format);
-
-std::tuple<torch::Tensor, int64_t> load_audio_fileobj(
-    py::object fileobj,
-    c10::optional<int64_t> frame_offset,
-    c10::optional<int64_t> num_frames,
-    c10::optional<bool> normalize,
-    c10::optional<bool> channels_first,
-    c10::optional<std::string> format);
-
-void save_audio_fileobj(
-    py::object fileobj,
-    torch::Tensor tensor,
-    int64_t sample_rate,
-    bool channels_first,
-    c10::optional<double> compression,
-    c10::optional<std::string> format,
-    c10::optional<std::string> encoding,
-    c10::optional<int64_t> bits_per_sample);
-
-#endif // TORCH_API_INCLUDE_EXTENSION_H
 
 } // namespace sox_io
 } // namespace torchaudio
